refactor: 패키지 구조 변경
- http , requestHandler 패키지 생성
