+++ src/main/java/webserver/http/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
